Start project over with lein figwheel new
I started the project over in order to gain
proper reloadable behaviour with figwheel.
This was not working properly with the
previous version.

Also, I managed to get a better display
option for the cells up and running - these
are now shown centred in square divs.

Whilst doing this, I discovered that
a (let) form nested in hiccup elements
and itself encasing hiccup elements
will only return the last element it
encases.

I really want to refactor the whole
shebang so that I can avoid such
hiccups (hehe).