Make number of cells around cellpointer variable
It is now possible to adjust the number of cells around the
cell at the cellpointer in the GUI with a slider. I mostly made
this in order to check out how it looks when I adjust the number
of cells. I might remove this later, along with the blinker-button
component, which is also only a development element.
But the real win was making the number of cells dynamic without
having them expand vertically - I fixed this using the 'flexbox'
css layout element. Thank you, flexbox!