Add , functionality and refactor step function
In brainfuck, the ',' character prompts the
user for input. The inputted character will
be converted to the corresponding ascii
number value and inserted into the current
cell.

In derpanet, this commit achieves this thusly;
the visual interpretation will halt upon
reaching a ',' character in the source code,
and wait for the user to write a character
and hit the button to read that character
into the current cell.

In addition, this commit also cleans
up some of the 'step' function -
but more cleanup is needed in order
for the function to be readable.