Move application to new reagent template
I was having problems figuring out how
to deploy the application on Heroku and
Redshift, and it seemed that it would be
simpler if the application had a backend.
Newer reagent templates utilize a ring
server, so I started up a new project,
moved the code and fixed up imports
and such so that the app would run.

I also fiddled with the styling a
little bit, adding line-height
specifications to the source-code
boxes (active and inactive),
since the browsers had changed
behaviour since I last saw derpanet,
and the inactive/active boxes would
show different line heights without
style specifcation.